Refactor: 대회일정 페이지 구조 변경 csr -> ssr
